Cambios verificacion 1

+++ backend/controllers/studentsController.php
@@ -32,18 +32,35 @@ function handleGet($conn)
 function handlePost($conn) 
 {
     $input = json_decode(file_get_contents("php://input"), true);
+    $email = $input['email'];
 
-    $result = createStudent($conn, $input['fullname'], $input['email'], $input['age']);
-    if ($result['inserted'] > 0) 
-    {
-        echo json_encode(["message" => "Estudiante agregado correctamente"]);
-    } 
-    else 
-    {
-        http_response_code(500);
-        echo json_encode(["error" => "No se pudo agregar"]);
+    $stmt = $conn->prepare("SELECT * FROM students WHERE email = ?");
+    $stmt->bind_param("s", $email);
+    $stmt->execute();
+    $check = $stmt->get_result();
+    $emailExiste = $check->num_rows > 0;
+
+    if ($emailExiste){
+        sendCodeMessage(409, "Email ya existente.");
     }
+    else
+    {
+        $result = createStudent($conn, $input['fullname'], $email, $input['age']);
+
+        if ($result['inserted'] > 0) 
+            {
+                echo json_encode(["message" => "Estudiante agregado correctamente"]);
+            } 
+            else 
+            {
+                http_response_code(500);
+                echo json_encode(["error" => "No se pudo agregar"]);
+            }
+    }
+
+    $stmt->close();
 }
+
 
 function handlePut($conn) 
 {
